debug free of vec storage
as mentioned in the issue #73 the drain iterator doesn't iterate
correctly on this unsafe inserted vector.
I don't know if it is expected and I will post the question on forums soon.

this commit contains a test that failed in past and an
implementation that pass this test.

basically it cleans the vector like it fills it by swapping value to drop
them and then set it len to 0