Added more documentation to world read / write methods.
Whilst there was already documentation explaining component / resource
IDs, it was easy to skim over and no reference was made by the
read_with_id and write_with_id methods. This made it seem like it
expected an entity ID to search for a component, whereas in reality it's
used for more complex component / resource setups.