Autoderive for Into/From saveload traits
`IntoSerialize` and `FromDeserialize` relied on `Component`
which was somewhat pointless. This wouldn't be a problem, but
due to how proc macros work, this is necessary to create something like
`\#[derive(IntoSerialize)]` because then we can ensure every field,
transitively, is `IntoSerialize`.

For the actual (De)`SerializeComponents` traits, this is changed
into `IntoSerialize+Component` and `FromDeserialize+Component` as
a bound to ensure correctness (it doesn't compile without it anyway).

We can then add `#[derive(Saveload)]` which auto-implements
`IntoSerialize`/`FromDeserialize` for most types (excluding those
with references/pointers, and ones with generics for now) by
automatically constructing a matching proxy "Data" type and converting
all the fields with a recursive call to `into`/`from`.