Fix typo substituting Position for Velocity
I believe the example intended to be `impl Component for Velocity` rather than `...for Position` as no `Position` struct was present in this example. I also added `#[derive(Default)]` for the `Gravity` struct as it was needed to have an example that would compile. I also gave the `run()` function a valid signature that would allow the example to compile.